requirements.txt
asgiref==3.7.2
Django==4.2.4
django-filter==23.2
django-pandas==0.6.6
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
numpy==1.25.2
packaging==23.1
pandas==2.0.3
# psycopg2==2.9.7
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
